+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 
 > NOTE: pywai follows the [semver](https://semver.org/) versioning standard.
 
-#### 0.0.1 (2024-11-11) **Latest**
+### 0.0.3 (2024-11-12) **Latest**
+
+- Add message_text and user_name to send_message
+
+### 0.0.2 (2024-11-12)
+
+- Add requirements
+- Rename to pywaai
+- Fixes to conversation_db
+- Add loguru and logfire optional dependencies
+
+
+#### 0.0.1 (2024-11-11)
 
 - Initial release